+++ daml-assistant/daml-helper/BUILD.bazel
@@ -11,12 +11,20 @@ da_haskell_binary(
         "//daml-assistant:daml-project-config",
     ],
     hazel_deps = [
+        "aeson",
+        "async",
         "base",
         "directory",
         "extra",
         "filepath",
+        "http-client",
+        "http-types",
+        "monad-loops",
+        "network",
         "optparse-applicative",
         "process",
+        "safe-exceptions",
+        "text",
     ],
     visibility = ["//visibility:public"],
 )
