+++ daml-assistant/src/DAML/Assistant/Tests.hs
@@ -7,10 +7,10 @@ module DAML.Assistant.Tests
     ( runTests
     ) where
 
-import DAML.Assistant.Types
-import DAML.Assistant.Consts
 import DAML.Assistant.Env
-import DAML.Assistant.Util
+import DAML.Project.Consts hiding (getDamlPath, getProjectPath)
+import DAML.Project.Types
+import DAML.Project.Util
 import System.Directory
 import System.Environment
 import System.FilePath
